фыв

+++ WebMSAPR/Data/models/Element.cs
@@ -3,12 +3,12 @@
 public class Element
 {
     public int Number{ get; set; }
-    public double Squre { get; set; }
-    public double Width { get; set; }
-    public double Length { get; set; }
+    public decimal Squre { get; set; }
+    public decimal Width { get; set; }
+    public decimal Length { get; set; }
     internal List<Tuple<Element, int>> AdjElement { get; set; } = new();
 
-    public Element(int N, int width, int length)
+    public Element(int N, decimal width, decimal length)
     {
         Number = N;
         Width = width==0?10:width;
